// shared/enums/ServerToSocketMsg.ts
export enum ServerToSocketMsg {
    CONNECTED = "CONNECTED",
    ENTITY_UPDATE = "ENTITY_UPDATE",
    START_ATTACK = "START_ATTACK",
    ATTACK = "ATTACK",
    ATTACK_RECEIVED = "ATTACK_RECEIVED",
    DISCONNECT = "DISCONNECT",
    CURRENT_ENTITIES = "CURRENT_ENTITIES",
    NEW_ENTITY = "NEW_ENTITY",
    ENTITY_DIED = "ENTITY_DIED",
    ENTITY_RESPAWNED = "ENTITY_RESPAWNED",
    ENTITY_POS_UPDATE = "ENTITY_POS_UPDATE",
    ENTITY_DIRECTION_UPDATE = "ENTITY_DIRECTION_UPDATE",
    ENTITY_SYNC = "ENTITY_SYNC",
    KNOCKBACK_RECEIVED = "KNOCKBACK_RECEIVED",
    ATTACK_RESULT = "ATTACK_RESULT"
}